fix: serveral bug when bump dependencies
