+++ .changeset/cold-timers-hug.md
@@ -1,0 +1,6 @@
+---
+'@ice/rspack-config': patch
+'@ice/app': patch
+---
+
+fix: css module hash and HMR
