+++ .changeset/fluffy-onions-travel.md
@@ -1,0 +1,6 @@
+---
+'@ice/webpack-config': patch
+'@ice/shared-config': patch
+---
+
+fix: optimize utils
