+++ packages/bundles/src/plugin-refresh.ts
@@ -1,0 +1,3 @@
+import RefreshPlugin from '@rspack/plugin-react-refresh';
+
+export default RefreshPlugin;
