+++ packages/shared-config/src/index.ts
@@ -4,6 +4,7 @@ import getCompilerPlugins from './getCompilerPlugins.js';
 import getDefineVars from './getDefineVars.js';
 import getPostcssOpts from './getPostcssOpts.js';
 import getCSSModuleLocalIdent from './getCSSModuleLocalIdent.js';
+import getAliasWithRoot from './getAlias.js';
 
 export {
   getCSSModuleLocalIdent,
@@ -15,4 +16,5 @@ export {
   getCompilerPlugins,
   getDefineVars,
   getPostcssOpts,
+  getAliasWithRoot,
 };
